backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic-settings>=2.5
sqlmodel>=0.0.22
python-multipart>=0.0.9
httpx>=0.27
tenacity>=8.3
redis>=5.0
rq>=1.16
orjson>=3.10
python-dateutil>=2.9
